requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-image-optimizer==1.0.1
gunicorn==20.1.0
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
optimize-images==1.4.0
pathspec==0.8.1
piexif==1.1.3
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
python-resize-image==1.1.19
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
tinify==1.5.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
yapf==0.31.0
